+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 - hyprland: active client empty on window close
 - dispose signal on Variable
 
+## Breaking Changes
+
+- Stream.is_muted corresponds to actual mute state
+
 # 1.8.0
 
 ## Features